renamed all the components and tests
