+++ src/components/icons/AsteriskIcon.js
