+++ src/components/icons/ClearAllIcon.js
