+++ src/components/icons/DeleteIcon.js
