+++ src/components/icons/DivideIcon.js
