+++ src/components/icons/EqualsIcon.js
