+++ src/components/icons/MinusIcon.js
