+++ src/components/icons/PercentIcon.js
