+++ src/components/icons/PeriodIcon.js
