+++ src/components/icons/PlusIcon.js
