+++ test/components/ButtonsGroup.test.js
@@ -2,7 +2,7 @@ import React from "react";
 import { expect, describe, it } from '@jest/globals';
 import renderer from "react-test-renderer";
 import Button from "../../src/components/Button";
-import ButtonsGroup from "../../src/components/Buttons-group";
+import ButtonsGroup from "../../src/components/ButtonsGroup";
 
 describe("pattern div component", () => {
